requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
boto3==1.34.144
pydantic==2.7.4
pytest==8.2.2
pytest-asyncio==0.23.7
uuid==1.30
cryptography==42.0.8
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1 